# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AuthLab API (DEV)
  version: "0.1.0"
  description: >
    Lab-only REST API for AuthLab.
    Cookie session is the primary auth for all endpoints.
    DEV bootstrap: POST /auth/session with Authorization: Bearer <DEV_API_KEY> (when DEV_MODE=true).
servers:
  - url: http://authlab:5000/api/v1
    description: Docker Compose network (service name authlab)
...
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Flask session cookie (set after auth/bootstrap).
    bearerAuth:
      type: http
      scheme: bearer
      description: DEV bootstrap only (POST /auth/session).
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: unauthorized
            message:
              type: string
              example: Login required
            details:
              nullable: true
    AuthSessionCreateResponse:
      type: object
      required: [user, csrf_token]
      properties:
        user:
          type: string
          example: admin
        csrf_token:
          type: string
          example: "7d1a0b5c...f3"
    AuthSessionStatusResponse:
      type: object
      required: [user]
      properties:
        user:
          type: string
          example: admin
...
paths:
  /auth/session:
    post:
      tags: [Auth]
      summary: Create/refresh session and return CSRF token (DEV bootstrap)
      description: >
        If session cookie already exists, returns user + csrf_token (ensuring token exists).
        Otherwise, in DEV_MODE you can bootstrap with Authorization: Bearer <DEV_API_KEY>.
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: OK
          headers:
            Set-Cookie:
              description: Flask session cookie
              schema:
                type: string
            Cache-Control:
              description: Response must not be cached.
              schema:
                type: string
              example: "no-store"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionCreateResponse"
        "401":
          description: Unauthorized (no valid session / invalid DEV bootstrap)
          headers:
            Cache-Control:
              description: Response must not be cached.
              schema:
                type: string
              example: "no-store"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          description: Too Many Requests (rate-limited)
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              schema:
                type: integer
                minimum: 1
            Cache-Control:
              description: Response must not be cached.
              schema:
                type: string
              example: "no-store"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    get:
      tags: [Auth]
      summary: Read current session status
      description: >
        Returns current user when session exists.
        If user exists but csrf_token is missing, returns 409 bootstrap_required.
      responses:
        "200":
          description: OK
          headers:
            Cache-Control:
              description: Response must not be cached.
              schema:
                type: string
              example: "no-store"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionStatusResponse"
        "401":
          description: Unauthorized (no valid session)
          headers:
            Cache-Control:
              description: Response must not be cached.
              schema:
                type: string
              example: "no-store"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Bootstrap required (csrf_token missing)
          headers:
            Cache-Control:
              description: Response must not be cached.
              schema:
                type: string
              example: "no-store"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error:
                  code: bootstrap_required
                  message: Call POST /auth/session first
...